requirements.txt
streamlit==1.31.0
pandas==2.1.1
requests==2.31.0
PyGithub==2.1.1
